add jks

+++ lib/features/search/provider/search_provider.dart
@@ -23,7 +23,7 @@ class SearchProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  int? range;
+  double? range;
   onSelectRange(v) {
     range = v;
     notifyListeners();
